0.8.9

+++ project/jimmer-sql-kotlin/src/test/kotlin/org/babyfish/jimmer/sql/kt/query/PagingTest.kt
@@ -197,6 +197,6 @@ class PagingTest : AbstractQueryTest() {
 
     data class Page<E>(
         val entities: List<E>,
-        val totalCount: Int
+        val totalCount: Long
     )
 }
